requirements.txt
torch==2.10.0
pandas==2.3.3
numpy==2.4.2
matplotlib==3.10.8
scikit-learn==1.8.0
shap==0.50.0
joblib==1.5.3
python-dotenv>=1.0.0
requests>=2.31.0
